+++ Lib/fontbakery/checks/opentype/head.py
@@ -1,7 +1,7 @@
 import fractions
 
 from fontbakery.callable import check
-from fontbakery.status import FAIL, PASS, WARN
+from fontbakery.status import FAIL, PASS, WARN, SKIP
 from fontbakery.message import Message
 from fontbakery.constants import NameID
 
@@ -190,6 +190,10 @@ def com_google_fonts_check_mac_style(font):
     """Checking head.macStyle value."""
     from fontbakery.utils import check_bit_entry
     from fontbakery.constants import MacStyle
+
+    if font.style is None:
+        yield SKIP, "Font style could not be determined."
+        return
 
     # Checking macStyle ITALIC bit:
     expected = "Italic" in font.style